+++ proto/ibc/lightclients/solomachine/v1/solomachine.proto
@@ -38,9 +38,9 @@ message Header {
       [(gogoproto.moretags) = "yaml:\"new_public_key\""];
 }
 
-// Evidence defines evidence of misbehaviour for a solo machine which consists
+// Misbehaviour defines evidence of misbehaviour for a solo machine which consists
 // of a sequence and two signatures over different messages at that sequence.
-message Evidence {
+message Misbehaviour {
   option (gogoproto.goproto_getters)  = false;
   option (gogoproto.goproto_stringer) = false;
   string           client_id = 1 [(gogoproto.moretags) = "yaml:\"client_id\""];
@@ -82,11 +82,11 @@ message MsgUpdateClient {
 }
 
 // MsgSubmitClientMisbehaviour defines an sdk.Msg type that supports submitting
-// arbitrary Evidence.
+// arbitrary Misbehaviour.
 message MsgSubmitClientMisbehaviour {
   option (gogoproto.goproto_getters) = false;
   bytes submitter                    = 1
       [(gogoproto.casttype) = "github.com/cosmos/cosmos-sdk/types.AccAddress"];
-  Evidence evidence = 2 [(gogoproto.nullable) = false];
+  Misbehaviour misbehaviour = 2 [(gogoproto.nullable) = false];
 }
