+++ proto/ibc/tendermint/tendermint.proto
@@ -68,9 +68,9 @@ message ConsensusState {
   ];
 }
 
-// Evidence is a wrapper over two conflicting Headers
-// that implements Evidence interface expected by ICS-02
-message Evidence {
+// Misbehaviour is a wrapper over two conflicting Headers
+// that implements Misbehaviour interface expected by ICS-02
+message Misbehaviour {
   option (gogoproto.goproto_getters)  = false;
   option (gogoproto.goproto_stringer) = false;
