+++ x/ibc/07-tendermint/types/misbehaviour.go
@@ -1,141 +1,179 @@
 package types
 
 import (
+	"math"
 	"time"
 
+	yaml "gopkg.in/yaml.v2"
+
+	"github.com/tendermint/tendermint/crypto/tmhash"
+	tmbytes "github.com/tendermint/tendermint/libs/bytes"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	tmtypes "github.com/tendermint/tendermint/types"
 
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	evidenceexported "github.com/cosmos/cosmos-sdk/x/evidence/exported"
 	clientexported "github.com/cosmos/cosmos-sdk/x/ibc/02-client/exported"
 	clienttypes "github.com/cosmos/cosmos-sdk/x/ibc/02-client/types"
+	host "github.com/cosmos/cosmos-sdk/x/ibc/24-host"
 )
 
-// CheckMisbehaviourAndUpdateState determines whether or not two conflicting
-// headers at the same height would have convinced the light client.
-//
-// NOTE: consensusState1 is the trusted consensus state that corresponds to the TrustedHeight
-// of misbehaviour.Header1
-// Similarly, consensusState2 is the trusted consensus state that corresponds
-// to misbehaviour.Header2
-func (cs ClientState) CheckMisbehaviourAndUpdateState(
-	ctx sdk.Context,
-	cdc codec.BinaryMarshaler,
-	clientStore sdk.KVStore,
-	misbehaviour clientexported.Misbehaviour,
-) (clientexported.ClientState, error) {
+var (
+	_ evidenceexported.Evidence   = Misbehaviour{}
+	_ clientexported.Misbehaviour = Misbehaviour{}
+)
 
-	// If client is already frozen at earlier height than evidence, return with error
-	if cs.IsFrozen() && cs.FrozenHeight <= uint64(misbehaviour.GetHeight()) {
-		return nil, sdkerrors.Wrapf(clienttypes.ErrInvalidEvidence,
-			"client is already frozen at earlier height %d than misbehaviour height %d", cs.FrozenHeight, misbehaviour.GetHeight())
+// NewMisbehaviour creates a new Misbehaviour instance.
+func NewMisbehaviour(clientID, chainID string, header1, header2 Header) *Misbehaviour {
+	return &Misbehaviour{
+		ClientId: clientID,
+		ChainId:  chainID,
+		Header1:  header1,
+		Header2:  header2,
 	}
 
-	tmEvidence, ok := misbehaviour.(*Evidence)
-	if !ok {
-		return nil, sdkerrors.Wrapf(clienttypes.ErrInvalidClientType, "expected type %T, got %T", misbehaviour, &Evidence{})
-	}
-
-	// Retrieve trusted consensus states for each Header in misbehaviour
-	// and unmarshal from clientStore
-
-	// Get consensus bytes from clientStore
-	tmConsensusState1, err := GetConsensusState(clientStore, cdc, tmEvidence.Header1.TrustedHeight)
-	if err != nil {
-		return nil, sdkerrors.Wrapf(err, "could not get trusted consensus state from clientStore for Header1 at TrustedHeight: %d", tmEvidence.Header1.TrustedHeight)
-	}
-
-	// Get consensus bytes from clientStore
-	tmConsensusState2, err := GetConsensusState(clientStore, cdc, tmEvidence.Header2.TrustedHeight)
-	if err != nil {
-		return nil, sdkerrors.Wrapf(err, "could not get trusted consensus state from clientStore for Header2 at TrustedHeight: %d", tmEvidence.Header2.TrustedHeight)
-	}
-
-	// calculate the age of the misbehaviour evidence
-	infractionHeight := tmEvidence.GetHeight()
-	infractionTime := tmEvidence.GetTime()
-	ageDuration := ctx.BlockTime().Sub(infractionTime)
-	ageBlocks := int64(cs.LatestHeight) - infractionHeight
-
-	// TODO: Retrieve consensusparams from client state and not context
-	// Issue #6516: https://github.com/cosmos/cosmos-sdk/issues/6516
-	consensusParams := ctx.ConsensusParams()
-
-	// Reject misbehaviour if the age is too old. Evidence is considered stale
-	// if the difference in time and number of blocks is greater than the allowed
-	// parameters defined.
-	//
-	// NOTE: The first condition is a safety check as the consensus params cannot
-	// be nil since the previous param values will be used in case they can't be
-	// retrieved. If they are not set during initialization, Tendermint will always
-	// use the default values.
-	if consensusParams != nil &&
-		consensusParams.Evidence != nil &&
-		(ageDuration > consensusParams.Evidence.MaxAgeDuration ||
-			ageBlocks > consensusParams.Evidence.MaxAgeNumBlocks) {
-		return nil, sdkerrors.Wrapf(clienttypes.ErrInvalidEvidence,
-			"age duration (%s) and age blocks (%d) are greater than max consensus params for duration (%s) and block (%d)",
-			ageDuration, ageBlocks, consensusParams.Evidence.MaxAgeDuration, consensusParams.Evidence.MaxAgeNumBlocks,
-		)
-	}
-
-	// Check the validity of the two conflicting headers against their respective
-	// trusted consensus states
-	// NOTE: header height and commitment root assertions are checked in
-	// evidence.ValidateBasic by the client keeper and msg.ValidateBasic
-	// by the base application.
-	if err := checkMisbehaviourHeader(
-		&cs, tmConsensusState1, tmEvidence.Header1, ctx.BlockTime(),
-	); err != nil {
-		return nil, sdkerrors.Wrap(err, "verifying Header1 in Evidence failed")
-	}
-	if err := checkMisbehaviourHeader(
-		&cs, tmConsensusState2, tmEvidence.Header2, ctx.BlockTime(),
-	); err != nil {
-		return nil, sdkerrors.Wrap(err, "verifying Header2 in Evidence failed")
-	}
-
-	cs.FrozenHeight = uint64(tmEvidence.GetHeight())
-	return &cs, nil
 }
 
-// checkMisbehaviourHeader checks that a Header in Misbehaviour is valid evidence given
-// a trusted ConsensusState
-func checkMisbehaviourHeader(
-	clientState *ClientState, consState *ConsensusState, header Header, currentTimestamp time.Time,
-) error {
+// ClientType is Tendermint light client
+func (misbehaviour Misbehaviour) ClientType() clientexported.ClientType {
+	return clientexported.Tendermint
+}
 
-	tmTrustedValset, err := tmtypes.ValidatorSetFromProto(header.TrustedValidators)
+// GetClientID returns the ID of the client that committed a misbehaviour.
+func (misbehaviour Misbehaviour) GetClientID() string {
+	return misbehaviour.ClientId
+}
+
+// Route implements Misbehaviour interface
+func (misbehaviour Misbehaviour) Route() string {
+	return clienttypes.SubModuleName
+}
+
+// Type implements Misbehaviour interface
+func (misbehaviour Misbehaviour) Type() string {
+	return "client_misbehaviour"
+}
+
+// String implements Misbehaviour interface
+func (misbehaviour Misbehaviour) String() string {
+	// FIXME: implement custom marshaller
+	bz, err := yaml.Marshal(misbehaviour)
 	if err != nil {
-		return sdkerrors.Wrap(err, "trusted validator set is not tendermint validator set type")
+		panic(err)
+	}
+	return string(bz)
+}
+
+// Hash implements Misbehaviour interface
+func (misbehaviour Misbehaviour) Hash() tmbytes.HexBytes {
+	bz := SubModuleCdc.MustMarshalBinaryBare(&misbehaviour)
+	return tmhash.Sum(bz)
+}
+
+// GetHeight returns the height at which misbehaviour occurred
+//
+// NOTE: assumes that evidence headers have the same height
+func (misbehaviour Misbehaviour) GetHeight() int64 {
+	return int64(math.Min(float64(misbehaviour.Header1.GetHeight()), float64(misbehaviour.Header2.GetHeight())))
+}
+
+// GetTime returns the timestamp at which misbehaviour occurred. It uses the
+// maximum value from both headers to prevent producing an invalid header outside
+// of the evidence age range.
+func (misbehaviour Misbehaviour) GetTime() time.Time {
+	minTime := int64(math.Max(float64(misbehaviour.Header1.GetTime().UnixNano()), float64(misbehaviour.Header2.GetTime().UnixNano())))
+	return time.Unix(0, minTime)
+}
+
+// ValidateBasic implements Misbehaviour interface
+func (misbehaviour Misbehaviour) ValidateBasic() error {
+	if misbehaviour.Header1.TrustedHeight == 0 {
+		return sdkerrors.Wrap(ErrInvalidHeaderHeight, "evidence Header1 must have non-zero trusted height")
+	}
+	if misbehaviour.Header2.TrustedHeight == 0 {
+		return sdkerrors.Wrap(ErrInvalidHeaderHeight, "evidence Header2 must have non-zero trusted height")
+	}
+	if misbehaviour.Header1.TrustedValidators == nil {
+		return sdkerrors.Wrap(ErrInvalidValidatorSet, "trusted validator set in Header1 cannot be empty")
+	}
+	if misbehaviour.Header2.TrustedValidators == nil {
+		return sdkerrors.Wrap(ErrInvalidValidatorSet, "trusted validator set in Header2 cannot be empty")
 	}
 
-	tmCommit, err := tmtypes.CommitFromProto(header.Commit)
+	if err := host.ClientIdentifierValidator(misbehaviour.ClientId); err != nil {
+		return sdkerrors.Wrap(err, "evidence client ID is invalid")
+	}
+
+	// ValidateBasic on both validators
+	if err := misbehaviour.Header1.ValidateBasic(misbehaviour.ChainId); err != nil {
+		return sdkerrors.Wrap(
+			clienttypes.ErrInvalidMisbehaviour,
+			sdkerrors.Wrap(err, "header 1 failed validation").Error(),
+		)
+	}
+	if err := misbehaviour.Header2.ValidateBasic(misbehaviour.ChainId); err != nil {
+		return sdkerrors.Wrap(
+			clienttypes.ErrInvalidMisbehaviour,
+			sdkerrors.Wrap(err, "header 2 failed validation").Error(),
+		)
+	}
+	// Ensure that Heights are the same
+	if misbehaviour.Header1.GetHeight() != misbehaviour.Header2.GetHeight() {
+		return sdkerrors.Wrapf(clienttypes.ErrInvalidMisbehaviour, "headers in evidence are on different heights (%d ≠ %d)", misbehaviour.Header1.GetHeight(), misbehaviour.Header2.GetHeight())
+	}
+
+	blockID1, err := tmtypes.BlockIDFromProto(&misbehaviour.Header1.SignedHeader.Commit.BlockID)
+	if err != nil {
+		return sdkerrors.Wrap(err, "invalid block ID from header 1 in evidence")
+	}
+	blockID2, err := tmtypes.BlockIDFromProto(&misbehaviour.Header2.SignedHeader.Commit.BlockID)
+	if err != nil {
+		return sdkerrors.Wrap(err, "invalid block ID from header 2 in evidence")
+	}
+
+	// Ensure that Commit Hashes are different
+	if blockID1.Equals(*blockID2) {
+		return sdkerrors.Wrap(clienttypes.ErrInvalidMisbehaviour, "headers blockIDs are not equal")
+	}
+	if err := ValidCommit(misbehaviour.ChainId, misbehaviour.Header1.Commit, misbehaviour.Header1.ValidatorSet); err != nil {
+		return err
+	}
+	if err := ValidCommit(misbehaviour.ChainId, misbehaviour.Header2.Commit, misbehaviour.Header2.ValidatorSet); err != nil {
+		return err
+	}
+	return nil
+}
+
+// ValidCommit checks if the given commit is a valid commit from the passed-in validatorset
+//
+// CommitToVoteSet will panic if the commit cannot be converted to a valid voteset given the validatorset
+// This implies that someone tried to submit evidence that wasn't actually committed by the validatorset
+// thus we should return an error here and reject the evidence rather than panicing.
+func ValidCommit(chainID string, commit *tmproto.Commit, valSet *tmproto.ValidatorSet) (err error) {
+	defer func() {
+		if r := recover(); r != nil {
+			err = sdkerrors.Wrapf(clienttypes.ErrInvalidMisbehaviour, "invalid commit: %v", r)
+		}
+	}()
+
+	tmCommit, err := tmtypes.CommitFromProto(commit)
 	if err != nil {
 		return sdkerrors.Wrap(err, "commit is not tendermint commit type")
 	}
-
-	// check the trusted fields for the header against ConsensusState
-	if err := checkTrustedHeader(header, consState); err != nil {
-		return err
+	tmValset, err := tmtypes.ValidatorSetFromProto(valSet)
+	if err != nil {
+		return sdkerrors.Wrap(err, "validator set is not tendermint validator set type")
 	}
 
-	// assert that the timestamp is not from more than an unbonding period ago
-	if currentTimestamp.Sub(consState.Timestamp) >= clientState.UnbondingPeriod {
-		return sdkerrors.Wrapf(
-			ErrUnbondingPeriodExpired,
-			"current timestamp minus the latest consensus state timestamp is greater than or equal to the unbonding period (%s >= %s)",
-			currentTimestamp.Sub(consState.Timestamp), clientState.UnbondingPeriod,
-		)
+	// Convert commits to vote-sets given the validator set so we can check if they both have 2/3 power
+	voteSet := tmtypes.CommitToVoteSet(chainID, tmCommit, tmValset)
+
+	blockID, ok := voteSet.TwoThirdsMajority()
+
+	// Check that ValidatorSet did indeed commit to blockID in Commit
+	if !ok || !blockID.Equals(tmCommit.BlockID) {
+		return sdkerrors.Wrap(clienttypes.ErrInvalidMisbehaviour, "validator set did not commit to header")
 	}
 
-	// - ValidatorSet must have 2/3 similarity with trusted FromValidatorSet
-	// - ValidatorSets on both headers are valid given the last trusted ValidatorSet
-	if err := tmTrustedValset.VerifyCommitLightTrusting(
-		clientState.GetChainID(), tmCommit, clientState.TrustLevel.ToTendermint(),
-	); err != nil {
-		return sdkerrors.Wrapf(clienttypes.ErrInvalidEvidence, "validator set in header has too much change from trusted validator set: %v", err)
-	}
 	return nil
 }
