+++ x/ibc/light-clients/solomachine/types/misbehaviour.go
@@ -1,62 +1,100 @@
 package types
 
 import (
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	"bytes"
+
+	yaml "gopkg.in/yaml.v2"
+
+	"github.com/tendermint/tendermint/crypto/tmhash"
+	tmbytes "github.com/tendermint/tendermint/libs/bytes"
+
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	evidenceexported "github.com/cosmos/cosmos-sdk/x/evidence/exported"
 	clientexported "github.com/cosmos/cosmos-sdk/x/ibc/02-client/exported"
 	clienttypes "github.com/cosmos/cosmos-sdk/x/ibc/02-client/types"
+	host "github.com/cosmos/cosmos-sdk/x/ibc/24-host"
 )
 
-// CheckMisbehaviourAndUpdateState determines whether or not the currently registered
-// public key signed over two different messages with the same sequence. If this is true
-// the client state is updated to a frozen status.
-func (cs ClientState) CheckMisbehaviourAndUpdateState(
-	ctx sdk.Context,
-	cdc codec.BinaryMarshaler,
-	clientStore sdk.KVStore,
-	misbehaviour clientexported.Misbehaviour,
-) (clientexported.ClientState, error) {
+var (
+	_ evidenceexported.Evidence   = (*Misbehaviour)(nil)
+	_ clientexported.Misbehaviour = (*Misbehaviour)(nil)
+)
 
-	evidence, ok := misbehaviour.(Evidence)
-	if !ok {
-		return nil, sdkerrors.Wrapf(
-			clienttypes.ErrInvalidClientType,
-			"evidence type %T, expected %T", misbehaviour, Evidence{},
-		)
-	}
-
-	if cs.IsFrozen() {
-		return nil, sdkerrors.Wrapf(clienttypes.ErrClientFrozen, "client is already frozen")
-	}
-
-	if err := checkMisbehaviour(cs, evidence); err != nil {
-		return nil, err
-	}
-
-	cs.FrozenSequence = uint64(evidence.GetHeight())
-	return cs, nil
+// ClientType is a Solo Machine light client.
+func (misbehaviour Misbehaviour) ClientType() clientexported.ClientType {
+	return clientexported.SoloMachine
 }
 
-// checkMisbehaviour checks if the currently registered public key has signed
-// over two different messages at the same sequence.
-// NOTE: a check that the evidence message data are not equal is done by
-// evidence.ValidateBasic which is called by the 02-client keeper.
-func checkMisbehaviour(clientState ClientState, evidence Evidence) error {
-	pubKey := clientState.ConsensusState.GetPubKey()
+// GetClientID returns the ID of the client that committed a misbehaviour.
+func (misbehaviour Misbehaviour) GetClientID() string {
+	return misbehaviour.ClientId
+}
 
-	data := EvidenceSignBytes(evidence.Sequence, evidence.SignatureOne.Data)
+// Route implements Evidence interface.
+func (misbehaviour Misbehaviour) Route() string {
+	return clienttypes.SubModuleName
+}
 
-	// check first signature
-	if err := VerifySignature(pubKey, data, evidence.SignatureOne.Signature); err != nil {
-		return sdkerrors.Wrap(err, "evidence signature one failed to be verified")
+// Type implements Evidence interface.
+func (misbehaviour Misbehaviour) Type() string {
+	return "client_misbehaviour"
+}
+
+// String implements Evidence interface.
+func (misbehaviour Misbehaviour) String() string {
+	out, _ := yaml.Marshal(misbehaviour)
+	return string(out)
+}
+
+// Hash implements Evidence interface
+func (misbehaviour Misbehaviour) Hash() tmbytes.HexBytes {
+	bz := SubModuleCdc.MustMarshalBinaryBare(&misbehaviour)
+	return tmhash.Sum(bz)
+}
+
+// GetHeight returns the sequence at which misbehaviour occurred.
+func (misbehaviour Misbehaviour) GetHeight() int64 {
+	return int64(misbehaviour.Sequence)
+}
+
+// ValidateBasic implements Evidence interface.
+func (misbehaviour Misbehaviour) ValidateBasic() error {
+	if err := host.ClientIdentifierValidator(misbehaviour.ClientId); err != nil {
+		return sdkerrors.Wrap(err, "invalid client identifier for solo machine")
 	}
 
-	data = EvidenceSignBytes(evidence.Sequence, evidence.SignatureTwo.Data)
+	if misbehaviour.Sequence == 0 {
+		return sdkerrors.Wrap(clienttypes.ErrInvalidMisbehaviour, "sequence cannot be 0")
+	}
 
-	// check second signature
-	if err := VerifySignature(pubKey, data, evidence.SignatureTwo.Signature); err != nil {
-		return sdkerrors.Wrap(err, "evidence signature two failed to be verified")
+	if err := misbehaviour.SignatureOne.ValidateBasic(); err != nil {
+		return sdkerrors.Wrap(err, "signature one failed basic validation")
+	}
+
+	if err := misbehaviour.SignatureTwo.ValidateBasic(); err != nil {
+		return sdkerrors.Wrap(err, "signature two failed basic validation")
+	}
+
+	// evidence signatures cannot be identical
+	if bytes.Equal(misbehaviour.SignatureOne.Signature, misbehaviour.SignatureTwo.Signature) {
+		return sdkerrors.Wrap(clienttypes.ErrInvalidMisbehaviour, "evidence signatures cannot be equal")
+	}
+
+	// message data signed cannot be identical
+	if bytes.Equal(misbehaviour.SignatureOne.Data, misbehaviour.SignatureTwo.Data) {
+		return sdkerrors.Wrap(clienttypes.ErrInvalidMisbehaviour, "evidence signature data must be signed over different messages")
+	}
+
+	return nil
+}
+
+// ValidateBasic ensures that the signature and data fields are non-empty.
+func (sd SignatureAndData) ValidateBasic() error {
+	if len(sd.Signature) == 0 {
+		return sdkerrors.Wrap(ErrInvalidSignatureAndData, "signature cannot be empty")
+	}
+	if len(sd.Data) == 0 {
+		return sdkerrors.Wrap(ErrInvalidSignatureAndData, "data for signature cannot be empty")
 	}
 
 	return nil
